src/server/django_rest/requirements.txt
Django==2.0.1
djangorestframework==3.7.3
opencv-python==3.4.0.12
django-cors-headers==2.0.0
keras==2.0.5
tensorflow==1.1.0
pandas==0.19.1
numpy==1.12.1
h5py==2.7.0
matplotlib
statistics
